+++ Tweaks/UiAdjustment/GearPositions.cs
@@ -1,5 +1,7 @@
-using FFXIVClientStructs.FFXIV.Component.GUI;
+using System;
+using System.Diagnostics.CodeAnalysis;
 using FFXIVClientStructs.FFXIV.Client.System.Memory;
+using FFXIVClientStructs.FFXIV.Component.GUI;
 using SimpleTweaksPlugin.Events;
 using SimpleTweaksPlugin.TweakSystem;
 using SimpleTweaksPlugin.Utility;
@@ -8,6 +10,7 @@ namespace SimpleTweaksPlugin.Tweaks.UiAdjustment;
 
 [Changelog("1.8.1.1", "Fixed widget display when using standard UI quality.")]
 [Changelog("1.9.0.0", "Improved gamepad navigation on Character window.")]
+[Changelog("1.9.1.0", "Further improved gamepad navigation on Character window.")]
 public unsafe class GearPositions : UiAdjustments.SubTweak {
     public override string Name => "Adjust Equipment Positions";
     public override string Description => "Repositions equipment positions in character menu and inspect to give a less gross layout.";
@@ -100,6 +103,30 @@ public unsafe class GearPositions : UiAdjustments.SubTweak {
         Piety = 21
     }
 
+    private enum CharacterProfileNode {
+        GrandCompany = 1,
+    }
+
+    [SuppressMessage("ReSharper", "InconsistentNaming")]
+    private enum CharacterClassNode {
+        WHM = 5,
+        SCH = 6,
+        AST = 7,
+        SGE = 8,
+        
+        BRD = 14,
+        MCH = 15,
+        DNC = 16,
+        BLM = 17,
+        SMN = 18,
+        RDM = 19,
+        BLU = 20,
+    }
+
+    private enum CharacterReputeNode {
+        Commend = 0,
+    }
+
     private int GetCollisionNodeIndex(AtkUnitBase* atkUnitBase) {
         for (var i = 0; i < atkUnitBase->CollisionNodeListCount; i++) {
             if (atkUnitBase->CollisionNodeList[i] == atkUnitBase->CursorTarget) {
@@ -109,21 +136,31 @@ public unsafe class GearPositions : UiAdjustments.SubTweak {
         return -1;
     }
     
+    private byte F(params Enum[] t){
+        byte FocusNode(AtkUnitBase* atkUnitBase, int node) {
+            if (node < 0 || node >= atkUnitBase->CollisionNodeListCount) return 1;
+            atkUnitBase->SetFocusNode(atkUnitBase->CollisionNodeList[node]);
+            atkUnitBase->CursorTarget = atkUnitBase->CollisionNodeList[node];
+            return 1;
+        }
 
-
-    private byte F(string addonName, int node) {
-        var atkUnitBase = Common.GetUnitBase(addonName);
-        if (atkUnitBase == null) return 1;
-        if (node < 0 || node >= atkUnitBase->CollisionNodeListCount) return 1;
-        atkUnitBase->SetFocusNode(atkUnitBase->CollisionNodeList[node]);
-        atkUnitBase->CursorTarget = atkUnitBase->CollisionNodeList[node];
+        foreach (var e in t) {
+            switch (e) {
+                case CharacterNode node when Common.GetUnitBase("Character", out var unitBase): return FocusNode(unitBase, (int)node);
+                case CharacterStatusNode node when Common.GetUnitBase("CharacterStatus", out var unitBase): return FocusNode(unitBase, (int)node);
+                case CharacterInspectNode node when Common.GetUnitBase("CharacterInspect", out var unitBase): return FocusNode(unitBase, (int)node);
+                case CharacterProfileNode node when Common.GetUnitBase("CharacterProfile", out var unitBase): return FocusNode(unitBase, (int)node);
+                case CharacterClassNode node when Common.GetUnitBase("CharacterClass", out var unitBase): return FocusNode(unitBase, (int)node);
+                case CharacterReputeNode node when Common.GetUnitBase("CharacterRepute", out var unitBase): return FocusNode(unitBase, (int)node);
+            }
+        }
         return 1;
     }
-    private byte F(CharacterInspectNode node) => F("CharacterInspect", (int)node);
-    private byte F(CharacterNode node) => F("Character", (int)node);
-    private byte F(CharacterStatusNode node) => F("CharacterStatus", (int)node);
+    
 
     private byte ControllerInputDetour(AtkUnitBase* atkUnitBase, Dir d, byte a3) {
+        var name = Common.ReadString(atkUnitBase->Name, 0x20);
+        SimpleLog.Log($"{name}, {GetCollisionNodeIndex(atkUnitBase)}, {d}");
         try {
             if (atkUnitBase == Common.GetUnitBase("CharacterStatus")) {
                 var currentSelectedNodeIndex = (CharacterStatusNode)GetCollisionNodeIndex(atkUnitBase);
@@ -148,19 +185,19 @@ public unsafe class GearPositions : UiAdjustments.SubTweak {
                     CharacterNode.SoulCrystal when d == Dir.Right => F(CharacterNode.OffHand),
                     CharacterNode.SoulCrystal when d == Dir.Down => F(CharacterNode.MainHand),
                     CharacterNode.SoulCrystal when d == Dir.Up => F(CharacterNode.GearSetList),
-                    CharacterNode.SoulCrystal when d == Dir.Left => F(CharacterStatusNode.Intelligence),
+                    CharacterNode.SoulCrystal when d == Dir.Left => F(CharacterStatusNode.Intelligence, CharacterProfileNode.GrandCompany, CharacterClassNode.WHM, CharacterReputeNode.Commend),
                     CharacterNode.MainHand when d == Dir.Up => F(CharacterNode.SoulCrystal),
-                    CharacterNode.MainHand when d == Dir.Left => F(CharacterStatusNode.Mind),
+                    CharacterNode.MainHand when d == Dir.Left => F(CharacterStatusNode.Mind, CharacterProfileNode.GrandCompany, CharacterClassNode.AST, CharacterReputeNode.Commend),
                     CharacterNode.Head when d == Dir.Right => F(CharacterNode.Ears),
-                    CharacterNode.Head when d == Dir.Left => F(CharacterStatusNode.Defense),
+                    CharacterNode.Head when d == Dir.Left => F(CharacterStatusNode.Defense, CharacterProfileNode.GrandCompany, CharacterClassNode.SGE, CharacterReputeNode.Commend),
                     CharacterNode.Body when d == Dir.Right => F(CharacterNode.Neck),
-                    CharacterNode.Body when d == Dir.Left => F(CharacterStatusNode.MagicDefense),
+                    CharacterNode.Body when d == Dir.Left => F(CharacterStatusNode.MagicDefense, CharacterProfileNode.GrandCompany, CharacterClassNode.BRD, CharacterReputeNode.Commend),
                     CharacterNode.Hands when d == Dir.Right => F(CharacterNode.Wrist),
-                    CharacterNode.Hands when d == Dir.Left => F(CharacterStatusNode.AttackMagicPotency),
+                    CharacterNode.Hands when d == Dir.Left => F(CharacterStatusNode.AttackMagicPotency, CharacterProfileNode.GrandCompany, CharacterClassNode.MCH, CharacterReputeNode.Commend),
                     CharacterNode.Legs when d == Dir.Right => F(CharacterNode.FingerRight),
-                    CharacterNode.Legs when d == Dir.Left => F(CharacterStatusNode.SpellSpeed),
+                    CharacterNode.Legs when d == Dir.Left => F(CharacterStatusNode.SpellSpeed, CharacterProfileNode.GrandCompany, CharacterClassNode.DNC, CharacterReputeNode.Commend),
                     CharacterNode.Feet when d == Dir.Right => F(CharacterNode.FingerLeft),
-                    CharacterNode.Feet when d == Dir.Left => F(CharacterStatusNode.Tenacity),
+                    CharacterNode.Feet when d == Dir.Left => F(CharacterStatusNode.Tenacity, CharacterProfileNode.GrandCompany, CharacterClassNode.BLM, CharacterReputeNode.Commend),
                     CharacterNode.OffHand when d == Dir.Left => F(CharacterNode.MainHand),
                     CharacterNode.Ears when d == Dir.Left => F(CharacterNode.Head),
                     CharacterNode.Neck when d == Dir.Left => F(CharacterNode.Body),
@@ -218,6 +255,40 @@ public unsafe class GearPositions : UiAdjustments.SubTweak {
                     CharacterInspectNode.FingerLeft when d == Dir.Down => F(CharacterInspectNode.ResetDisplay),
                     CharacterInspectNode.ResetDisplay when d == Dir.Right => F(CharacterInspectNode.FingerLeft),
                     CharacterInspectNode.ResetDisplay when d == Dir.Up => F(CharacterInspectNode.FingerLeft),
+                    _ => addonControllerInputHook.Original(atkUnitBase, d, a3)
+                };
+            }
+
+            if (atkUnitBase == Common.GetUnitBase("CharacterClass")) {
+                var currentSelectedNodeIndex = (CharacterClassNode)GetCollisionNodeIndex(atkUnitBase);
+                return currentSelectedNodeIndex switch {
+                    CharacterClassNode.WHM when d == Dir.Right => F(CharacterNode.SoulCrystal),
+                    CharacterClassNode.SCH when d == Dir.Right => F(CharacterNode.MainHand),
+                    CharacterClassNode.AST when d == Dir.Right => F(CharacterNode.MainHand),
+                    CharacterClassNode.SGE when d == Dir.Right => F(CharacterNode.Head),
+                    CharacterClassNode.BRD when d == Dir.Right => F(CharacterNode.Body),
+                    CharacterClassNode.MCH when d == Dir.Right => F(CharacterNode.Hands),
+                    CharacterClassNode.DNC when d == Dir.Right => F(CharacterNode.Legs),
+                    CharacterClassNode.BLM when d == Dir.Right => F(CharacterNode.Feet),
+                    CharacterClassNode.SMN when d == Dir.Right => F(CharacterNode.Feet),
+                    CharacterClassNode.RDM when d == Dir.Right => F(CharacterNode.Feet),
+                    CharacterClassNode.BLU when d == Dir.Right => F(CharacterNode.Feet),
+                    _ => addonControllerInputHook.Original(atkUnitBase, d, a3)
+                };
+            }
+            
+            if (atkUnitBase == Common.GetUnitBase("CharacterProfile")) {
+                var currentSelectedNodeIndex = (CharacterProfileNode)GetCollisionNodeIndex(atkUnitBase);
+                return currentSelectedNodeIndex switch {
+                    CharacterProfileNode.GrandCompany when d == Dir.Right => F(CharacterNode.SoulCrystal),
+                    _ => addonControllerInputHook.Original(atkUnitBase, d, a3)
+                };
+            }
+            
+            if (atkUnitBase == Common.GetUnitBase("CharacterRepute")) {
+                var currentSelectedNodeIndex = (CharacterReputeNode)GetCollisionNodeIndex(atkUnitBase);
+                return currentSelectedNodeIndex switch {
+                    CharacterReputeNode.Commend when d == Dir.Right => F(CharacterNode.SoulCrystal),
                     _ => addonControllerInputHook.Original(atkUnitBase, d, a3)
                 };
             }